% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpikeSolver.R
\name{BayesSpikeSolver}
\alias{BayesSpikeSolver}
\title{Designate Bayes Spike as the TReNA Solver and Solve}
\usage{
BayesSpikeSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\value{
A Solver class object with Bayes Spike as the solver
}
\description{
Designate Bayes Spike as the TReNA Solver and Solve
}
