% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateFilter.R
\name{CandidateFilter-Class}
\alias{CandidateFilter-Class}
\alias{CandidateFilter}
\title{Class CandidateFilter}
\usage{
CandidateFilter(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the CandidateFilter object should print output}
}
\value{
An object of the Candidate filter class
}
\examples{
# Create an empty candidate filter
candidate.filter <- CandidateFilter(mtx.assay = matrix(), quiet=TRUE)
}
