% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{closeDatabaseConnections,FootprintFinder-method}
\alias{closeDatabaseConnections,FootprintFinder-method}
\title{Close a Footprint Database Connection}
\usage{
\S4method{closeDatabaseConnections}{FootprintFinder}(obj)
}
\arguments{
\item{obj}{An object of class FootprintFinder}
}
\description{
Close a Footprint Database Connection
}
