% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleSolver.R
\docType{class}
\name{EnsembleSolver-class}
\alias{EnsembleSolver-class}
\alias{.EnsembleSolver}
\title{An S4 class to represent an Ensemble Solver}
\description{
An S4 class to represent an Ensemble Solver
}
