% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filter.R
\docType{class}
\name{Filter-class}
\alias{Filter-class}
\alias{.Filter}
\title{An S4 class to represent a filter}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\description{
An S4 class to represent a filter
}
