% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFilter.R
\docType{class}
\name{FootprintFilter-class}
\alias{FootprintFilter-class}
\alias{.FootprintFilter}
\title{Filter based on gene footprints}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\description{
Filter based on gene footprints
}
