% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{class}
\name{FootprintFinder-class}
\alias{FootprintFinder-class}
\alias{.FootprintFinder}
\alias{FootprintFinder}
\alias{FootprintFinder-class}
\alias{FootprintFinder}
\alias{closeDatabaseConnections,FootprintFinder-method}
\title{Class FootprintFinder}
\usage{
FootprintFinder(genome.database.uri, project.database.uri, quiet = TRUE)

\S4method{closeDatabaseConnections}{FootprintFinder}(obj)
}
\arguments{
\item{genome.database.uri}{The address of a genome database for use in filtering. This database
must contain the tables "gtf" and "motifsgenes" at a minimum. The URI format is as follows:
"dbtype://host/database" (e.g. "postgres://localhost/genomedb")}

\item{project.database.uri}{The address of a project database for use in filtering. This database
must contain the tables "regions" and "hits" at a minimum. The URI format is as follows:
"dbtype://host/database" (e.g. "postgres://localhost/projectdb")}

\item{quiet}{A logical denoting whether or not the FootprintFinder object should print output}

\item{obj}{An object of class FootprintFinder}
}
\value{
An object of the FootprintFinder class that can reduce a list of genes to a subset prior
to forming a TReNA object

An object of the FootprintFinder class

Closes the specified database connection
}
\description{
The FootprintFinder class is designed to query 2 supplied footprint databases (a genome database
and a project database) for supplied genes or regions. Within the TReNA package, the
FootprintFinder class is mainly used by the FootprintFilter class, but the FootprintFinder class
offers more flexibility in constructing queries.

This method takes a FootprintFinder object and closes connections to the footprint databases
if they are currently open.
}
\section{Slots}{

\describe{
\item{\code{genome.db}}{The address of a genome database for use in filtering}

\item{\code{project.db}}{The address of a project database for use in filtering}

\item{\code{quiet}}{A logical argument denoting whether the FootprintFinder object should behave quietly}
}}

\seealso{
\code{\link{FootprintFilter}}
}
