% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{class}
\name{FootprintFinder-class}
\alias{FootprintFinder-class}
\alias{.FootprintFinder}
\alias{FootprintFinder-class}
\alias{FootprintFinder}
\title{Filter genes based on footprints}
\usage{
FootprintFinder(genome.database.uri, project.database.uri, quiet = TRUE)
}
\value{
An object of the FootprintFinder class that can reduce a list of genes to a subset prior
to forming a TReNA object
}
\description{

}
\section{Slots}{

\describe{
\item{\code{genome.db}}{The address of a genome database for use in filtering}

\item{\code{project.db}}{The address of a project database for use in filtering}

\item{\code{quiet}}{A logical argument denoting whether the FootprintFinder object should behave quietly}
}}

