% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\docType{methods}
\name{getAssayData-methods}
\alias{getAssayData-methods}
\alias{getAssayData}
\title{Get Assay Data from Solver}
\usage{
\S4method{getAssayData}{Solver}(obj)
}
\arguments{
\item{obj}{An object of class Solver}
}
\value{
The assay matrix of gene expression data associated with a Solver object
}
\description{
Retrieve the assay matrix of gene expression data from a Solver object
}
\examples{

# Create a Solver object using the included Alzheimer's data and retrieve the matrix
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
solver <- Solver(mtx.sub)
mtx <- getAssayData(solver)
}
