% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\docType{methods}
\name{getAssayData,Solver-method}
\alias{getAssayData,Solver-method}
\title{Get Assay Data from Solver}
\usage{
\S4method{getAssayData}{Solver}(obj)
}
\value{
The assay matrix of gene expression data associated with a Solver object
}
\description{
Get Assay Data from Solver
}
