% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{getAssayData}
\alias{getAssayData}
\title{Retrieve the assay matrix of gene expression data from a Solver object}
\usage{
getAssayData(obj)
}
\arguments{
\item{obj}{An object of class Solver}
}
\value{
The assay matrix of gene expression data associated with a Solver object
}
\description{
Retrieve the assay matrix of gene expression data from a Solver object
}
