% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFilter.R
\docType{methods}
\name{getCandidates,FootprintFilter-method}
\alias{getCandidates,FootprintFilter-method}
\alias{getCandidates-FootprintFilter}
\title{Get candidate genes using the variance filter}
\usage{
\S4method{getCandidates}{FootprintFilter}(obj, target.gene, genome.db.uri,
  project.db.uri, size.upstream = 1000, size.downstream = 1000)
}
\arguments{
\item{obj}{An object of class VarianceFilter}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{genome.db.uri}{A connection to a genome database containing footprint information}

\item{project.db.uri}{A connection to a project database containing footprint information}

\item{size.upstream}{An integer denoting the distance upstream of the target gene to look for footprints}

\item{size.downstream}{An integer denoting the distance downstream of the target gene to look for footprints}
}
\value{
A vector containing all genes with variances less than the target gene
}
\description{
Get candidate genes using the variance filter
}
