% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullFilter.R
\docType{methods}
\name{getCandidates,NullFilter-method}
\alias{getCandidates,NullFilter-method}
\alias{getCandidates-NullFilter}
\title{Get candidate genes using the null filter}
\usage{
\S4method{getCandidates}{NullFilter}(obj)
}
\arguments{
\item{obj}{An object of class NullFilter}
}
\value{
A vector containing all genes in the assay matrix
}
\description{
Get candidate genes using the null filter
}
