% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenChromFilter.R
\docType{methods}
\name{getCandidates,OpenChromFilter-method}
\alias{getCandidates,OpenChromFilter-method}
\alias{getCandidates-OpenChromFilter}
\title{Get candidate genes using the open chromatin filter}
\usage{
tfs <- getCandidates(obj, chromasome, start, end)
}
\arguments{
\item{obj}{An object of class FootprintFilter}

\item{start}{An integer denoting the starting point of the region of interest}

\item{end}{An integer denoting the ending point of the region of interest}

\item{chromasome}{A chromasome of interest that contains the regions to be used for filtering}
}
\value{
A vector containing all genes with motifs in the supplied region
}
\description{
Get candidate genes using the open chromatin filter
}
\examples{

# Use open chromatin filter for MEF2C to filter candidates
# in the included Alzheimer's dataset
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
openchrom.filter <- OpenChromFilter(mtx.assay = mtx.sub)

}
\seealso{
\code{\link{OpenChromFilter}}

Other getCandidate Methods: \code{\link{getCandidates,FootprintFilter-method}},
  \code{\link{getCandidates,NullFilter-method}},
  \code{\link{getCandidates,VarianceFilter-method}}
}
