% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateFilter.R
\name{getCandidates}
\alias{getCandidates}
\alias{CandidateFilter}
\title{Get candidate genes using the selected filter}
\usage{
CandidateFilter(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\value{
A vector containing all genes selected as candidates by the filter
}
\description{
Get candidate genes using the selected filter
}
