% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{getChromLoc,FootprintFinder-method}
\alias{getChromLoc,FootprintFinder-method}
\alias{getChromLoc}
\title{Get Chromasome Location}
\usage{
\S4method{getChromLoc}{FootprintFinder}(obj, name, biotype = "protein_coding",
  moleculetype = "gene")
}
\arguments{
\item{obj}{An object of class FootprintFinder}

\item{name}{A gene name or ID}

\item{biotype}{A type of biological unit (default="protein_coding")}

\item{moleculetype}{A type of molecule (default="gene")}
}
\value{
A dataframe containing the results of a database query pertaining to the specified name,
biotype, and molecule type. This dataframe contains the following columns: gene_id, gene_name,
chr, start, endpos, strand
}
\description{
Using the gtf table in the genome database contained in a FootprintFinder object, get the locations
of chromasomes with the specified gene name, biological unit type, and molecule type
}
