% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTReNA.R
\name{getDnaseEnhancerRegions}
\alias{getDnaseEnhancerRegions}
\title{Get DNase Enhancer Regions}
\usage{
getDnaseEnhancerRegions()
}
\value{
A table of enhancers where the method is DNase-DNase
}
\description{
Connect to the SQL database of hg38 enhancers and select only those regions where method = DNase-DNase
}
