% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTReNA.R
\name{getFootprintsForTF}
\alias{getFootprintsForTF}
\title{Get Footprints for a TF}
\usage{
getFootprintsForTF(obj, tf)
}
\arguments{
\item{tf}{A transcription factor}
}
\description{
Get all footprints from the project database for a specified transcription factor
}
