% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{getFootprintsInRegion,FootprintFinder-method}
\alias{getFootprintsInRegion,FootprintFinder-method}
\alias{getFootprintsInRegion}
\title{Get Footprints in a Region}
\usage{
\S4method{getFootprintsInRegion}{FootprintFinder}(obj, chromosome, start, end)
}
\arguments{
\item{obj}{An object of class FootprintFinder}

\item{chromosome}{The name of the chromosome of interest}

\item{start}{An integer denoting the start of the desired region}

\item{end}{An integer denoting the end of the desired region}
}
\value{
A dataframe containing all footprints for the specified region
}
\description{
Using the regions and hits tables inside the project database specified by the FootprintFinder
object, return the location, chromasome, starting position, and ending positions of all footprints
for the specified region.
}
