% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTReNA.R
\name{getGenelistFromGtf}
\alias{getGenelistFromGtf}
\title{Get Gene List from Gtf}
\usage{
getGenelistFromGtf(genome.db.uri, project.db.uri, biotype = "protein_coding",
  moleculetype = "gene", use_gene_ids = T)
}
\arguments{
\item{genome.db.uri}{A web address to a genome database}

\item{project.db.uri}{A web address to a project database}

\item{biotype}{A string specifying the biotype of interest (default = "protein_coding") **What are the options?}

\item{moleculetype}{A string specifying the moleculetype of interest (default = "gene") **What are the options?}

\item{use_gene_ids}{A logical character indicating whether or not to use gene IDs as opposed to gene names (default = T)}
}
\value{
A list of genes resulting from the query
}
\description{
Query the SQL gtf table to get a list of genes matching the supplied "biotype" and "moleculetype"
}
