% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTReNA.R
\name{getGenePromoterRegions}
\alias{getGenePromoterRegions}
\title{Get Gene Promoter Regions}
\usage{
getGenePromoterRegions(obj, genelist, size.upstream = 10000,
  size.downstream = 10000)
}
\arguments{
\item{obj}{An object of class FootprintFilter(??)}

\item{genelist}{A gene list}

\item{size.upstream}{Number of basepairs to grab upstream of each gene (default = 10000)}

\item{size.downstream}{Number of basepairs to grab downstream of each gene (default = 10000)}
}
\description{
Get the promoter regions for a list of genes. Region sizes can be tuned by specifying
the positions upstream and downstream of the gene that bound the region.
}
