% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{getGtfGeneBioTypes,FootprintFinder-method}
\alias{getGtfGeneBioTypes,FootprintFinder-method}
\alias{getGtfGeneBioTypes}
\title{Get the List of Biotypes}
\usage{
\S4method{getGtfGeneBioTypes}{FootprintFinder}(obj)
}
\arguments{
\item{obj}{An object of class FootprintFinder}
}
\value{
A sorted list of the types of biological units contained in the gtf table of the genome
database.
}
\description{
Using the gtf table in the genome database contained in a FootprintFinder object, get the list of
different types of biological units (biotypes) contained in the table.
}
