% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTReNA.R
\name{getHiCEnhancerRegions}
\alias{getHiCEnhancerRegions}
\title{Get Hi-C Enhancer Regions}
\usage{
getHiCEnhancerRegions()
}
\value{
A table of enhancers where the method is Hi-C
}
\description{
Connect to the SQL database of hg38 enhancers and select only those regions where method = Hi-C
}
