% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{getPromoterRegionsAllGenes-methods}
\alias{getPromoterRegionsAllGenes-methods}
\alias{getPromoterRegionsAllGenes}
\title{Get Promoter Regions for All Genes}
\usage{
\S4method{getPromoterRegionsAllGenes}{FootprintFinder}(obj,
  size.upstream = 10000, size.downstream = 10000, use_gene_ids = T)
}
\arguments{
\item{obj}{An object of class FootprintFinder}

\item{size.upstream}{An integer denoting the distance upstream of each gene's transcription start
site to include in the promoter region (default = 1000)}

\item{size.downstream}{An integer denoting the distance downstream of each gene's transcription start
site to include in the promoter region (default = 1000)}

\item{use_gene_ids}{A binary indicating whether to return gene IDs or gene names (default = T)}
}
\value{
A GRanges object containing the promoter regions for all genes
}
\description{
Using the gtf table inside the genome database specified by the FootprintFinder object, return the
promoter regions for every protein-coding gene in the database.
}
