% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpikeSolver.R
\docType{methods}
\name{getSolverName,BayesSpikeSolver-method}
\alias{getSolverName,BayesSpikeSolver-method}
\title{Get Bayes Spike Solver name}
\usage{
\S4method{getSolverName}{BayesSpikeSolver}(obj)
}
\arguments{
\item{obj}{An object of the class BayesSpikeSolver}
}
\value{
"BayesSpikeSolver"
}
\description{
Get Bayes Spike Solver name
}
