% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnsembleSolver.R
\docType{methods}
\name{getSolverName,EnsembleSolver-method}
\alias{getSolverName,EnsembleSolver-method}
\title{Get Ensemble Solver name}
\usage{
\S4method{getSolverName}{EnsembleSolver}(obj)
}
\arguments{
\item{obj}{An object of class EnsembleSolver}
}
\value{
"EnsembleSolver"
}
\description{
Get Ensemble Solver name
}
\examples{
solver <- EnsembleSolver()
getSolverName(solver)
}
