% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoPVSolver.R
\docType{methods}
\name{getSolverName,LassoPVSolver-method}
\alias{getSolverName,LassoPVSolver-method}
\title{Get LassoPV Solver name}
\usage{
\S4method{getSolverName}{LassoPVSolver}(obj)
}
\arguments{
\item{obj}{An object of class LassoPVSolver}
}
\value{
"LassoPVSolver"
}
\description{
Get LassoPV Solver name
}
\examples{
solver <- LassoPVSolver()
getSolverName(solver)
}
