% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoSolver.R
\docType{methods}
\name{getSolverName,LassoSolver-method}
\alias{getSolverName,LassoSolver-method}
\title{Get Lasso Solver name}
\usage{
\S4method{getSolverName}{LassoSolver}(obj)
}
\value{
"LassoSolver"
}
\description{
Get Lasso Solver name
}
