% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSolver.R
\docType{methods}
\name{getSolverName,RandomForestSolver-method}
\alias{getSolverName,RandomForestSolver-method}
\title{Get RandomForest Solver Name}
\usage{
\S4method{getSolverName}{RandomForestSolver}(obj)
}
\arguments{
\item{obj}{An object of class RandomForestSolver}
}
\value{
"RandomForestSolver"
}
\description{
Get RandomForest Solver Name
}
\examples{
solver <- RandomForestSolver()
getSolverName(solver)
}
