% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeSolver.R
\docType{methods}
\name{getSolverName,RidgeSolver-method}
\alias{getSolverName,RidgeSolver-method}
\title{Get Ridge Solver name}
\usage{
\S4method{getSolverName}{RidgeSolver}(obj)
}
\arguments{
\item{obj}{An object of class RidgeSolver}
}
\value{
"RidgeSolver"
}
\description{
Get Ridge Solver name
}
\examples{
solver <- RidgeSolver()
getSolverName(solver)
}
