% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpearmanSolver.R
\docType{methods}
\name{getSolverName,SpearmanSolver-method}
\alias{getSolverName,SpearmanSolver-method}
\title{Get Spearman Solver Name}
\usage{
\S4method{getSolverName}{SpearmanSolver}(obj)
}
\value{
"SpearmanSolver"
}
\description{
Get Spearman Solver Name
}
\examples{
solver <- SpearmanSolver()
getSolverName(solver)
}
