% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\docType{methods}
\name{getSolverName,SqrtLassoSolver-method}
\alias{getSolverName,SqrtLassoSolver-method}
\title{Get SqrtLasso Solver name}
\usage{
\S4method{getSolverName}{SqrtLassoSolver}(obj)
}
\arguments{
\item{obj}{An object of class SqrtLassoSolver}
}
\value{
"SqrtLassoSolver"
}
\description{
Get SqrtLasso Solver name
}
\examples{
solver <- SqrtLassoSolver()
getSolverName(solver)
}
