% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TReNA.R
\docType{methods}
\name{getSolverName,TReNA-method}
\alias{getSolverName,TReNA-method}
\alias{getSolverName}
\title{Get the Solver Name from a TReNA Object}
\usage{
\S4method{getSolverName}{TReNA}(obj)
}
\arguments{
\item{obj}{An object of class TReNA}
}
\value{
A string giving the name of the solver type for the given Solver object
}
\description{
Retrieve the name of the solver specified in a TReNA object
}
\examples{

# Create a LassoSolver object using the included Alzheimer's data and retrieve the solver name
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
solver <- TReNA(mtx.sub, solver = "lasso")
mtx <- getSolverName(solver)
}
