% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TReNA.R
\docType{methods}
\name{getSolverObject,TReNA-method}
\alias{getSolverObject,TReNA-method}
\alias{getSolverObject}
\title{Get the Solver Object from a TReNA Object}
\usage{
\S4method{getSolverObject}{TReNA}(obj)
}
\arguments{
\item{obj}{An object of class TReNA}
}
\value{
The Solver object contained by the given TReNA object
}
\description{
Retrieve the name of the solver specified in a TReNA object
}
\examples{

# Create a LassoSolver object using the included Alzheimer's data and retrieve the solver object
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
solver <- TReNA(mtx.sub, solver = "lasso")
mtx <- getSolverObject(solver)
}
