% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTReNA.R
\name{getTfbsCountsInEnhancers}
\alias{getTfbsCountsInEnhancers}
\title{Get Transcription Factor Binding Site Counts in Enhancers}
\usage{
getTfbsCountsInEnhancers(tflist = NULL, genelist = NULL,
  enhancertype = "Hi-C", verbose = 2, cores = 5, genome.db.uri,
  project.db.uri, use_gene_ids = T, biotype = "protein_coding",
  moleculetype = "gene")
}
\description{
Get Transcription Factor Binding Site Counts in Enhancers
}
