% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTReNA.R
\name{getTfbsCountsPerPromoter}
\alias{getTfbsCountsPerPromoter}
\title{Get Transcription Factor Binding Site Counts Per Promoter}
\usage{
getTfbsCountsPerPromoter(obj, tflist, promoter_regions, verbose = 1)
}
\arguments{
\item{obj}{An object of class FootprintFilter}

\item{tflist}{A list of transcription factors}

\item{verbose}{A logical indicating whether the function should produce verbose output (default = 1)}

\item{promotor_regions}{A list of promoter regions}
}
\value{
A numeric vector containing transcription factor binding site counts per promoter
}
\description{
For a FootprintFilter object, get the transcription factor binding site counts for each promoter
}
