% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTReNA.R
\name{getTfbsCountsPerPromoterMC}
\alias{getTfbsCountsPerPromoterMC}
\title{Get Transcription Factor Binding Site Counts Per Promoter with Clustering}
\usage{
getTfbsCountsPerPromoterMC(tflist, promoter_regions, verbose = 1, cores = 5,
  genome.db.uri, project.db.uri)
}
\arguments{
\item{tflist}{A list of transcription factors}

\item{promoter_regions}{A list of promoter regions}

\item{verbose}{A logical indicating whether the function should produce verbose output (default = 1)}

\item{cores}{The number of computational cores to be devoted to the calculation (default = 5)}

\item{genome.db.uri}{A web address to the genome database}

\item{project.db.uri}{A web address to the project database}
}
\value{
A numeric vector containing transcription factor binding site counts per gene
}
\description{
Get the transription factor binding site counts for each promoter and cluster
}
