% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HumanDNAseClusterFilter.R
\docType{class}
\name{HumanDNAseClusterFilter-class}
\alias{HumanDNAseClusterFilter-class}
\alias{.HumanDNAseClusterFilter}
\alias{HumanDNAseClusterFilter}
\alias{HumanDNAseClusterFilter}
\title{Create a HumanDNAseClusterFilter objects}
\usage{
HumanDNAseClusterFilter(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the filter should print output}
}
\description{
An HumanDNAseClusterFilter object allows for filtering based on a supplied chromosome and region. Using its
associated \code{getCandidates} method, a chromosome, and starting/ending locations for a region,
an HumanDNAseClusterFilter object can be used to filter a list of possible transcription factors to those
that match motifs within the supplied region
}
\examples{
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
openchrom.filter <- HumanDNAseClusterFilter(mtx.assay = mtx.sub)
}
\seealso{
\code{\link{getCandidates-HumanDNAseClusterFilter}}

Other Filtering Objects: \code{\link{FootprintFilter-class}},
  \code{\link{NullFilter-class}},
  \code{\link{VarianceFilter-class}}
}
