% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullFilter.R
\docType{class}
\name{NullFilter-class}
\alias{NullFilter-class}
\alias{.NullFilter}
\title{Apply a null filter}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\value{
An object of class NullFilter
}
\description{
Apply a null filter
}
