% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NullFilter.R
\name{NullFilter}
\alias{NullFilter}
\title{Define an object of class Null Filter}
\usage{
NullFilter(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\value{
An object of the Null Filter class
}
\description{
Define an object of class Null Filter
}
