% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenChromFilter.R
\docType{class}
\name{OpenChromFilter-class}
\alias{OpenChromFilter-class}
\alias{.OpenChromFilter}
\alias{OpenChromFilter}
\alias{OpenChromFilter}
\title{Create a OpenChromFilter object}
\usage{
OpenChromFilter(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the filter should print output}
}
\description{
An OpenChromFilter object allows for filtering based on a supplied chromasome and region. Using its
associated \code{getCandidates} method, a chromasome, and starting/ending locations for a region, 
an OpenChromFilter object can be used to filter a list of possible transcription factors to those
that match motifs within the supplied region
}
\examples{
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
openchrom.filter <- OpenChromFilter(mtx.assay = mtx.sub)
}
\seealso{
\code{\link{getCandidates-OpenChromFilter}}

Other Filtering Objects: \code{\link{FootprintFilter-class}},
  \code{\link{NullFilter-class}},
  \code{\link{VarianceFilter-class}}
}
