% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpikeSolver.R
\docType{methods}
\name{predictFromModel,BayesSpikeSolver-method}
\alias{predictFromModel,BayesSpikeSolver-method}
\alias{predictFromModel.BayesSpikeSolver}
\title{Predict expression using a Bayes Spike Model}
\usage{
\S4method{predictFromModel}{BayesSpikeSolver}(obj, model, tfs, test.samples)
}
\arguments{
\item{model}{A Bayes Spike model generated using \code{\\link{trainModel.BayesSpikeSolver}} that relates the transcription
factors to the target gene of interest}

\item{tfs}{A list of transcription factor predictors matching those used in the Bayes Spike model}

\item{test.samples}{A matrix of gene expression levels for the supplied transcription factors}
}
\value{
Predictions of the target gene's expression level for each sample in the test sample set
}
\description{
Predict expression using a Bayes Spike Model
}
\seealso{
\code{\\link{trainModel.BayesSpikeSolver}}
}
