% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoSolver.R
\docType{methods}
\name{predictFromModel,LassoSolver-method}
\alias{predictFromModel,LassoSolver-method}
\alias{predictFromModel.LassoSolver}
\title{Predict expression using a LASSO Model}
\usage{
\S4method{predictFromModel}{LassoSolver}(obj, model, tfs, test.samples)
}
\arguments{
\item{model}{A LASSO model generated using \code{\\link{trainModel}} that relates the transcription
factors to the target gene of interest}

\item{tfs}{A list of transcription factor predictors matching those used in the LASSO model}

\item{test.samples}{A matrix of gene expression levels for the supplied transcription factors}
}
\value{
Predictions of the target gene's expression level for each sample in the test sample set
}
\description{
Predict expression using a LASSO Model
}
\seealso{
\code{\\link{trainModel}}
}
