% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoPVSolver.R
\docType{methods}
\name{rescalePredictorWeights,LassoPVSolver-method}
\alias{rescalePredictorWeights,LassoPVSolver-method}
\alias{rescalePredictorWeights.LassoPVSolver}
\title{Rescale LASSO P-Value Predictor Weights}
\usage{
\S4method{rescalePredictorWeights}{LassoPVSolver}(obj, rawValue.min,
  rawValue.max, rawValues)
}
\arguments{
\item{obj}{An object of class LassoPVSolver}

\item{rawValue.min}{The minimum value of the raw expression values}

\item{rawValue.max}{The maximum value of the raw expression values}

\item{rawValues}{A matrix of raw expression values}
}
\value{
A matrix of the raw values re-scaled using the minimum and maximum values
}
\description{
Rescale LASSO P-Value Predictor Weights
}
