% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoSolver.R
\docType{methods}
\name{rescalePredictorWeights,LassoSolver-method}
\alias{rescalePredictorWeights,LassoSolver-method}
\alias{rescalePredictorWeights.LassoSolver}
\title{Rescale LASSO Predictor Weights}
\usage{
\S4method{rescalePredictorWeights}{LassoSolver}(obj, rawValue.min, rawValue.max,
  rawValues)
}
\arguments{
\item{obj}{An object of class LassoSolver}

\item{rawValue.min}{The minimum value of the raw expression values}

\item{rawValue.max}{The maximum value of the raw expression values}

\item{rawValues}{A matrix of raw expression values}
}
\value{
A matrix of the raw values re-scaled using the minimum and maximum values
}
\description{
Rescale LASSO Predictor Weights
}
