% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeSolver.R
\docType{methods}
\name{rescalePredictorWeights,RidgeSolver-method}
\alias{rescalePredictorWeights,RidgeSolver-method}
\alias{rescalePredictorWeights.RidgeSolver}
\title{Rescale Ridge Regression Predictor Weights}
\usage{
\S4method{rescalePredictorWeights}{RidgeSolver}(obj, rawValue.min, rawValue.max,
  rawValues)
}
\arguments{
\item{obj}{An object of class RidgeSolver}

\item{rawValue.min}{The minimum value of the raw expression values}

\item{rawValue.max}{The maximum value of the raw expression values}

\item{rawValues}{A matrix of raw expression values}
}
\value{
A matrix of the raw values re-scaled using the minimum and maximum values
}
\description{
Rescale Ridge Regression Predictor Weights
}
