% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\docType{methods}
\name{rescalePredictorWeights,SqrtLassoSolver-method}
\alias{rescalePredictorWeights,SqrtLassoSolver-method}
\alias{rescalePredictorWeights.SqrtLassoSolver}
\title{Rescale Square Root LASSO Predictor Weights}
\usage{
\S4method{rescalePredictorWeights}{SqrtLassoSolver}(obj, rawValue.min,
  rawValue.max, rawValues)
}
\arguments{
\item{obj}{An object of class SqrtLassoSolver}

\item{rawValue.min}{The minimum value of the raw expression values}

\item{rawValue.max}{The maximum value of the raw expression values}

\item{rawValues}{A matrix of raw expression values}
}
\value{
A matrix of the raw values re-scaled using the minimum and maximum values
}
\description{
Rescale Square Root LASSO Predictor Weights
}
