% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpikeSolver.R
\name{rescalePredictorWeights}
\alias{rescalePredictorWeights}
\alias{rescalePredictorWeights.BayesSpikeSolver}
\title{Rescale Bayes Spike Predictor Weights}
\usage{
rescalePredictorWeights(obj, rawValue.min, rawValue.max, rawValues)
}
\arguments{
\item{obj}{An object of class BayesSpikeSolver}

\item{rawValue.min}{The minimum value of the raw expression values}

\item{rawValue.max}{The maximum value of the raw expression values}

\item{rawValues}{A matrix of raw expression values}
}
\value{
A matrix of the raw values re-scaled using the minimum and maximum values
}
\description{
Rescale Bayes Spike Predictor Weights
}
