% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpikeSolver.R
\docType{methods}
\name{run,BayesSpikeSolver-method}
\alias{run,BayesSpikeSolver-method}
\alias{run.BayesSpikeSolver}
\title{Run the Bayes Spike Solver}
\usage{
\S4method{run}{BayesSpikeSolver}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs), extraArgs = list()))
}
\arguments{
\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}

\item{extraArgs}{Modifiers to the Bayes Spike solver}
}
\value{
A data frame containing the coefficients relating the target gene to each transcription factor, plus other fit parameters.
}
\description{
Given a TReNA object with Bayes Spike as the solver, use the \code{\link{vbsr}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
}
\seealso{
\code{\link{vbsr}}
}
