% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoPVSolver.R
\docType{methods}
\name{run,LassoPVSolver-method}
\alias{run,LassoPVSolver-method}
\alias{run.LassoPVSolver}
\title{Run the LASSO P-Value Solver}
\usage{
\S4method{run}{LassoPVSolver}(obj, target.gene, tfs, extraArgs = list())
}
\arguments{
\item{obj}{An object of class LassoPVSolver}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{extraArgs}{Modifiers to the LASSO P-Value solver}
}
\value{
A data frame containing the p-values for each transcription factor pertaining to the target gene
plus the Pearson correlations between each transcription factor and the target gene.
}
\description{
Given a TReNA object with LASSO P-Value as the solver, use the \code{\link{lassopv}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
}
\examples{

}
\seealso{
\code{\link{lassopv}}
}
