% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoSolver.R
\docType{methods}
\name{run,LassoSolver-method}
\alias{run,LassoSolver-method}
\alias{run.LassoSolver}
\title{Run the LASSO Solver}
\usage{
\S4method{run}{LassoSolver}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs)), extraArgs = list())
}
\arguments{
\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}

\item{extraArgs}{Modifiers to the LASSO solver}
}
\value{
A data frame containing the coefficients relating the target gene to each transcription factor, plus other fit parameters.
}
\description{
Given a TReNA object with LASSO as the solver, use the \code{\link{glmnet}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
}
\examples{

}
\seealso{
\code{\link{glmnet}}
}
