% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearsonSolver.R
\docType{methods}
\name{run,PearsonSolver-method}
\alias{run,PearsonSolver-method}
\alias{run.PearsonSolver}
\title{Run the Pearson Solver}
\usage{
\S4method{run}{PearsonSolver}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs), extraArgs = list()))
}
\arguments{
\item{obj}{An object of class PearsonSolver}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}
}
\value{
fit The set of Pearson Correlation Coefficients between each transcription factor and the target gene.
}
\description{
Given a TReNA object with Pearson as the solver, use the \code{\link{cor}} function to
estimate coefficients for each transcription factor as a perdictor of the target gene's expression level
}
\seealso{
\code{\link{cor}}
}
