% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSolver.R
\docType{methods}
\name{run,RandomForestSolver-method}
\alias{run,RandomForestSolver-method}
\alias{run.RandomForestSolver}
\title{Run the Random Forest Solver}
\usage{
\S4method{run}{RandomForestSolver}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs), extraArgs = list()))
}
\arguments{
\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}

\item{extraArgs}{Modifiers to the Random Forest solver}
}
\value{
A list containing various parameters of the Random Forest fit.
}
\description{
Run the Random Forest Solver
}
