% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpearmanSolver.R
\docType{methods}
\name{run,SpearmanSolver-method}
\alias{run,SpearmanSolver-method}
\alias{run.SpearmanSolver}
\title{Run the Spearman Solver}
\usage{
\S4method{run}{SpearmanSolver}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs), extraArgs = list()))
}
\arguments{
\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}
}
\value{
fit The set of Spearman Correlation Coefficients between each transcription factor and the target gene.
}
\description{
Given a TReNA object with Spearman as the solver, use the \code{\link{cor}} function with
\code{method = "spearman"} to esimate coefficients for each transcription factor as a predictor of the target
gene's expression level
}
\seealso{
\code{\link{cor}}
}
