% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\docType{methods}
\name{run,SqrtLassoSolver-method}
\alias{run,SqrtLassoSolver-method}
\alias{run.SqrtLassoSolver}
\title{Run the Square Root LASSO Solver}
\usage{
\S4method{run}{SqrtLassoSolver}(obj, target.gene, tfs, extraArgs = list())
}
\arguments{
\item{obj}{An object of class SqrtLassoSolver}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{extraArgs}{Modifiers to the Square RootLASSO solver}
}
\value{
A data frame containing the coefficients relating the target gene to each transcription factor, plus other fit parameters.
}
\description{
Given a TReNA object with Square Root LASSO as the solver, use the \code{\link{slim}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
}
\examples{

}
\seealso{
\code{\link{slim}}
}
