% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSolver.R
\docType{methods}
\name{show,RandomForestSolver-method}
\alias{show,RandomForestSolver-method}
\title{Show a RandomForestSolver object}
\usage{
\S4method{show}{RandomForestSolver}(object)
}
\arguments{
\item{object}{An object of class RandomForestSolver}
}
\value{
The names of the top 10 candidate regulators, plus statistics on the assay matrix, the
target gene, and the total number of candidate regulators
}
\description{
Show a RandomForestSolver object
}
\examples{
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
tfs <- setdiff(rownames(mtx.sub),"MEF2C")
rf.solver <- RandomForestSolver(mtx.sub, "MEF2C", tfs)
show(rf.solver)
}
