% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TReNA.R
\docType{methods}
\name{solve,TReNA-method}
\alias{solve,TReNA-method}
\title{Solve the TReNA object}
\usage{
\S4method{solve}{TReNA}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs)), extraArgs = list())
}
\arguments{
\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}

\item{extraArgs}{Modifiers to the Bayes Spike solver}
}
\value{
A data frame containing coefficients relating the target gene to each transcription factor
}
\description{
Solve the TReNA object
}
