% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearsonSolver.R
\docType{methods}
\name{run,PearsonSolver-method}
\alias{run,PearsonSolver-method}
\alias{run.PearsonSolver}
\alias{solve.Pearson}
\title{Run the Pearson Solver}
\usage{
trena <- TReNA(mtx.assay, solver = "pearson")
tbl.out <- solve(trena, target.gene, tfs)
}
\arguments{
\item{obj}{An object of class PearsonSolver}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}
}
\value{
fit The set of Pearson Correlation Coefficients between each transcription factor and the target gene.
}
\description{
Given a TReNA object with Pearson as the solver, use the \code{\link{cor}} function to
estimate coefficients for each transcription factor as a perdictor of the target gene's expression level
}
\examples{
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
trena <- TReNA(mtx.assay = mtx.sub, solver = "pearson")
target.gene <- "MEF2C"
tfs <- setdiff(rownames(mtx.sub), target.gene)
tbl <- solve(trena, target.gene, tfs)
}
\seealso{
\code{\link{cor}}

Other solver methods: \code{\link{run,BayesSpikeSolver-method}},
  \code{\link{run,LassoPVSolver-method}},
  \code{\link{run,LassoSolver-method}},
  \code{\link{run,RandomForestSolver-method}},
  \code{\link{run,RidgeSolver-method}},
  \code{\link{run,SpearmanSolver-method}},
  \code{\link{run,SqrtLassoSolver-method}}
}
