% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\docType{class}
\name{Solver-class}
\alias{Solver-class}
\alias{.Solver}
\alias{Solver}
\title{Solver Class}
\usage{
Solver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical indicating whether or not the Solver object should print output}
}
\value{
An object of the Solver class
}
\description{
The Solver class is a generic class that governs the different solvers available in TReNA. A
Solver class object is constructed during creation of a TReNA object and resides within the
TReNA object. It is rarely called by itself; rather, interaction with a particular solver object
is achieved using the \code{\link{solve}} method on a TReNA object.

Define an object of class Solver
}
\section{Slots}{

\describe{
\item{\code{mtx.assay}}{An assay matrix of gene expression data}

\item{\code{quiet}}{A logical element indicating whether the Solver object should print output}
}}

\examples{
mtx <- matrix(rnorm(10000), nrow = 100)
solver <- Solver(mtx)

}
\seealso{
\code{\link{TReNA}}, \code{\link{solve}}
}
