% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\name{Solver}
\alias{Solver}
\title{Define an object of class Solver}
\usage{
Solver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\value{
An object of the Solver class
}
\description{
Define an object of class Solver
}
