% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpearmanSolver.R
\docType{methods}
\name{SpearmanSolver}
\alias{SpearmanSolver}
\alias{run,SpearmanSolver-method}
\alias{run.SpearmanSolver}
\title{Create a Solver class object using Spearman correlation coefficients as the solver}
\usage{
SpearmanSolver(mtx.assay = matrix(), quiet = TRUE)

\S4method{run}{SpearmanSolver}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs), extraArgs = list()))
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}

\item{obj}{An object of class SpearmanSolver}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}
}
\value{
A Solver class object with Spearman correlation coefficients as the solver

fit The set of Spearman Correlation Coefficients between each transcription factor and the target gene.
}
\description{
Create a Solver class object using Spearman correlation coefficients as the solver

Given a TReNA object with Spearman as the solver, use the \code{\link{cor}} function with
\code{method = "spearman"} to esimate coefficients for each transcription factor as a predictor of the target
gene's expression level
}
\examples{
solver <- SpearmanSolver()
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
trena <- TReNA(mtx.assay = mtx.sub, solver = "pearson")
target.gene <- "APOE"
tfs <- setdiff(rownames(mtx.sub), target.gene)
tbl <- solve(trena, target.gene, tfs)
}
\seealso{
\code{\link{cor}}
}
