% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\docType{methods}
\name{SqrtLassoSolver}
\alias{SqrtLassoSolver}
\alias{run,SqrtLassoSolver-method}
\alias{run.SqrtLassoSolver}
\title{Create a Solver class object using the Square Root LASSO solver}
\usage{
SqrtLassoSolver(mtx.assay = matrix(), quiet = TRUE)

\S4method{run}{SqrtLassoSolver}(obj, target.gene, tfs, extraArgs = list())
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}

\item{obj}{An object of class SqrtLassoSolver}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{extraArgs}{Modifiers to the Square Root LASSO solver}
}
\value{
A Solver class object with Square Root LASSO as the solver

A data frame containing the coefficients relating the target gene to each transcription factor, plus other fit parameters.
}
\description{
Create a Solver class object using the Square Root LASSO solver

Given a TReNA object with Square Root LASSO as the solver, use the \code{\link{slim}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
}
\examples{
solver <- SqrtLassoSolver()
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
trena <- TReNA(mtx.assay = mtx.sub, solver = "sqrtlasso")
target.gene <- "APOE"
tfs <- setdiff(rownames(mtx.sub), target.gene)
tbl <- solve(trena, target.gene, tfs)
}
\seealso{
\code{\link{slim}}
}
