% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesSpikeSolver.R
\docType{methods}
\name{trainModel,BayesSpikeSolver-method}
\alias{trainModel,BayesSpikeSolver-method}
\alias{trainModel.BayesSpikeSolver}
\title{Train a model using Bayes Spike}
\usage{
\S4method{trainModel}{BayesSpikeSolver}(obj, target.gene, tfs, training.samples,
  tf.weights = rep(1, length(tfs)))
}
\arguments{
\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{training.samples}{***These aren't currently used!}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}
}
\value{
A model relating the target gene (response) to the transcription factors (predictors)
}
\description{
Train a model using Bayes Spike
}
