\name{TReNA}
\alias{TReNA}
\alias{TReNA-package}
\title{TReNA:  Fit transcriptional regulatory networks using transcription factor binding sites and lasso regression}
\description{
  Methods for reconstructing transcriptional regulatory networks, especially in species
  for which genome-wide TF binding site information is available.
  }

\keyword{networks}


