% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarianceFilter.R
\docType{class}
\name{VarianceFilter-class}
\alias{VarianceFilter-class}
\alias{.VarianceFilter}
\title{Filter based on gene expression variance}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\description{
Filter based on gene expression variance
}
