% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateFilter.R
\docType{class}
\name{CandidateFilter-class}
\alias{CandidateFilter-class}
\alias{.CandidateFilter}
\alias{CandidateFilter}
\alias{CandidateFilter}
\title{Class CandidateFilter}
\usage{
CandidateFilter(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the CandidateFilter object should print output}
}
\value{
An object of the CandidateFilter class

An object of the Candidate filter class
}
\description{
Class CandidateFilter

A CandidateFilter is an S4 class to represent a gene candidate filter. These filters can employ a variety of methods
to reduce the number of transcription factors used as predictors for solving a TReNA object.
}
\section{Slots}{

\describe{
\item{\code{mtx.assay}}{An assay matrix of gene expression data}

\item{\code{quiet}}{A logical denoting whether or not the CandidateFilter object should print output}
}}

\examples{
# Create an empty candidate filter
candidate.filter <- CandidateFilter(mtx.assay = matrix(), quiet=TRUE)
}
\seealso{
\code{\link{FootprintFilter}} \code{\link{VarianceFilter}} \code{\link{NullFilter}}
}
