% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CandidateFilter.R
\docType{class}
\name{CandidateFilter-class}
\alias{CandidateFilter-class}
\alias{.CandidateFilter}
\title{An S4 class to represent a gene candidate filter}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\description{
An S4 class to represent a gene candidate filter
}
