% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{class}
\name{FootprintFinder-class}
\alias{FootprintFinder-class}
\alias{.FootprintFinder}
\title{Filter genes based on footprints}
\arguments{
\item{genome.database.uri}{The address of a genome database for use in filtering}

\item{project.database.uri}{The address of a project database for use in filtering}
}
\value{
An object of the Filter class that reduces a list of genes to a subset prior to forming a TReNA object
}
\description{
Filter genes based on footprints
}
