% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenChromFilter.R
\docType{methods}
\name{getCandidates,OpenChromFilter-method}
\alias{getCandidates,OpenChromFilter-method}
\alias{getCandidates-OpenChromFilter}
\title{Get candidate genes using the open chromatin filter}
\usage{
\S4method{getCandidates}{OpenChromFilter}(obj, extraArgs)
}
\arguments{
\item{obj}{An object of class FootprintFilter}

\item{extraArgs}{\itemize{
\item{"chromosome" A chromosome of interest that contains the regions to be used for filtering}
\item{"start" An integer denoting the starting point of the region of interest}
\item{"end" An integer denoting the ending point of the region of interest}
}}
}
\value{
A vector containing all genes with motifs in the supplied region
}
\description{
Get candidate genes using the open chromatin filter
}
\examples{

# Use open chromatin filter for MEF2C to filter candidates
# in the included Alzheimer's dataset
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
openchrom.filter <- OpenChromFilter(mtx.assay = mtx.sub)
tfs <- getCandidates(openchrom.filter, extraArgs = list("chromosome" = "chr5",
"start" = 888936629, "end" = 898936629))
}
\seealso{
\code{\link{OpenChromFilter}}

Other getCandidate Methods: \code{\link{getCandidates,CandidateFilter-method}},
  \code{\link{getCandidates,FootprintFilter-method}},
  \code{\link{getCandidates,NullFilter-method}},
  \code{\link{getCandidates,VarianceFilter-method}}
}
