% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarianceFilter.R
\docType{methods}
\name{getCandidates,VarianceFilter-method}
\alias{getCandidates,VarianceFilter-method}
\alias{getCandidates-VarianceFilter}
\title{Get candidate genes using the variance filter}
\usage{
\S4method{getCandidates}{VarianceFilter}(obj, target.gene)
}
\arguments{
\item{target.gene}{A designated target gene that should be part of the mtx.assay data}
}
\value{
A vector containing all genes with variances less than the target gene
}
\description{
Get candidate genes using the variance filter
}
