% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearsonSolver.R
\docType{methods}
\name{getSolverName,PearsonSolver-method}
\alias{getSolverName,PearsonSolver-method}
\title{Get Pearson Solver name}
\usage{
\S4method{getSolverName}{PearsonSolver}(obj)
}
\value{
"PearsonSolver
}
\description{
Get Pearson Solver name
}
