% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSolver.R
\docType{methods}
\name{getSolverName,RandomForestSolver-method}
\alias{getSolverName,RandomForestSolver-method}
\title{Get Random Forest Solver name}
\usage{
\S4method{getSolverName}{RandomForestSolver}(obj)
}
\value{
"RandomForestSolver"
}
\description{
Get Random Forest Solver name
}
