% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\docType{methods}
\name{getSolverName,Solver-method}
\alias{getSolverName,Solver-method}
\alias{getSolverName}
\title{Get the Solver Name from a Solver Object}
\usage{
\S4method{getSolverName}{Solver}(obj)
}
\arguments{
\item{obj}{An object of class Solver}
}
\value{
A string giving the name of the solver type for the given Solver object
}
\description{
Retrieve the name of the solver specified in a solver object
}
\examples{

# Create a LassoSolver object using the included Alzheimer's data and retrieve the solver name
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
solver <- LassoSolver(mtx.sub)
mtx <- getSolverName(solver)
}
