% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoPVSolver.R
\name{LassoPVSolver}
\alias{LassoPVSolver}
\title{Create a Solver class object using the LASSO P-Value solver}
\usage{
LassoPVSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\value{
A Solver class object with LASSO P-Value as the solver
}
\description{
Create a Solver class object using the LASSO P-Value solver
}
\examples{
solver <- LassoPVSolver()
}
