% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoPVSolver.R
\docType{methods}
\name{LassoPVSolver}
\alias{LassoPVSolver}
\alias{run,LassoPVSolver-method}
\alias{run.LassoPVSolver}
\title{Create a Solver class object using the LASSO P-Value solver}
\usage{
LassoPVSolver(mtx.assay = matrix(), quiet = TRUE)

\S4method{run}{LassoPVSolver}(obj, target.gene, tfs, extraArgs = list())
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}

\item{obj}{An object of class LassoPVSolver}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{extraArgs}{Modifiers to the LASSO P-Value solver}
}
\value{
A Solver class object with LASSO P-Value as the solver

A data frame containing the p-values for each transcription factor pertaining to the target gene
plus the Pearson correlations between each transcription factor and the target gene.
}
\description{
Create a Solver class object using the LASSO P-Value solver

Given a TReNA object with LASSO P-Value as the solver, use the \code{\link{lassopv}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
}
\examples{
solver <- LassoPVSolver()
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
trena <- TReNA(mtx.assay = mtx.sub, solver = "lassopv")
target.gene <- "APOE"
tfs <- setdiff(rownames(mtx.sub), target.gene)
tbl <- solve(trena, target.gene, tfs)
}
\seealso{
\code{\link{lassopv}}
}
