% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LassoSolver.R
\name{LassoSolver}
\alias{LassoSolver}
\title{Create a Solver class object using the LASSO solver}
\usage{
LassoSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with LASSO as the solver
}
\description{
Create a Solver class object using the LASSO solver
}
\examples{
solver <- LassoSolver()
}
