% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FootprintFinder.R
\docType{methods}
\name{mapMotifsToTFsMergeIntoTable,FootprintFinder-method}
\alias{mapMotifsToTFsMergeIntoTable,FootprintFinder-method}
\alias{mapMotifsToTFsMergeIntoTable}
\title{Map Motifs to Transcription Factors and Merge into a Table}
\usage{
\S4method{mapMotifsToTFsMergeIntoTable}{FootprintFinder}(obj, tbl)
}
\arguments{
\item{obj}{An object of class FootprintFinder}

\item{tbl}{A dataframe of footprints, generally obtained using \code{\link{getFootprintsInRegion}}
or \code{\link{getFootprintsForGene}}}
}
\value{
A data frame containing the motifs from the supplied footprints table and the transcription
factors they map to
}
\description{
Using the motifsgenes table inside the genome database specified by the FootprintFinder object,
return a table mapping each motif to transcription factors
}
\seealso{
\code{\link{getFootprintsInRegion}}, \code{\link{getFootprintsForGene}}
}
