% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PearsonSolver.R
\name{PearsonSolver}
\alias{PearsonSolver}
\title{Create a Solver class object using  Pearson correlation coefficients as the solver}
\usage{
PearsonSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\description{
Create a Solver class object using  Pearson correlation coefficients as the solver
}
