% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomForestSolver.R
\docType{class}
\name{RandomForestSolver-class}
\alias{RandomForestSolver-class}
\alias{.RandomForestSolver}
\title{An S4 class to represent a Random Forest solver}
\description{
An S4 class to represent a Random Forest solver
}
