% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RidgeSolver.R
\docType{methods}
\name{RidgeSolver}
\alias{RidgeSolver}
\alias{run,RidgeSolver-method}
\alias{run.RidgeSolver}
\title{Create a Solver class object using the Ridge Regression solver}
\usage{
RidgeSolver(mtx.assay = matrix(), quiet = TRUE)

\S4method{run}{RidgeSolver}(obj, target.gene, tfs, tf.weights = rep(1,
  length(tfs)), extraArgs = list())
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}

\item{obj}{An object of class RidgeSolver}

\item{target.gene}{A designated target gene that should be part of the mtx.assay data}

\item{tfs}{The designated set of transcription factors that could be associated with the target gene.}

\item{tf.weights}{A set of weights on the transcription factors (default = rep(1, length(tfs)))}

\item{extraArgs}{Modifiers to the Ridge Regression solver}
}
\value{
A Solver class object with Ridge Regression as the solver

A data frame containing the coefficients relating the target gene to each transcription factor, plus other fit parameters.
}
\description{
Create a Solver class object using the Ridge Regression solver

Given a TReNA object with Ridge Regression as the solver, use the \code{\link{glmnet}} function to estimate coefficients
for each transcription factor as a predictor of the target gene's expression level.
}
\examples{
solver <- RidgeSolver()
# Load included Alzheimer's data, create a TReNA object with Bayes Spike as solver, and solve
load(system.file(package="TReNA", "extdata/ampAD.154genes.mef2cTFs.278samples.RData"))
trena <- TReNA(mtx.assay = mtx.sub, solver = "ridge")
target.gene <- "APOE"
tfs <- setdiff(rownames(mtx.sub), target.gene)
tbl <- solve(trena, target.gene, tfs)
}
\seealso{
\code{\link{glmnet}}
}
