% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Solver.R
\docType{class}
\name{Solver-class}
\alias{Solver-class}
\alias{.Solver}
\title{An S4 class to represent a solver}
\description{
An S4 class to represent a solver
}
\section{Slots}{

\describe{
\item{\code{mtx.assay}}{An assay matrix of gene expression data}

\item{\code{quiet}}{A logical element indicating whether the solver should produce output}

\item{\code{state}}{Environment}
}}

