% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpearmanSolver.R
\name{SpearmanSolver}
\alias{SpearmanSolver}
\title{Create a Solver class object using Spearman correlation coefficients as the solver}
\usage{
SpearmanSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{quiet}{A logical denoting whether or not the solver should print output}
}
\value{
A Solver class object with Spearman correlation coefficients as the solver
}
\description{
Create a Solver class object using Spearman correlation coefficients as the solver
}
\examples{
solver <- SpearmanSolver()
}
