% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SqrtLassoSolver.R
\name{SqrtLassoSolver}
\alias{SqrtLassoSolver}
\title{Create a Solver class object using the Square Root LASSO solver}
\usage{
SqrtLassoSolver(mtx.assay = matrix(), quiet = TRUE)
}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}
}
\value{
A Solver class object with Square Root LASSO as the solver
}
\description{
Create a Solver class object using the Square Root LASSO solver
}
\examples{
solver <- SqrtLassoSolver()
}
