% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TReNA.R
\docType{class}
\name{TReNA-class}
\alias{TReNA-class}
\alias{.TReNA}
\alias{TReNA}
\alias{TReNA-class}
\alias{TReNA}
\title{Class TReNA}
\usage{
TReNA(mtx.assay = matrix(), solver = "lasso", quiet = TRUE)
}
\value{
An object of the TReNA class
}
\description{
Class \code{TReNA} defines a TReNA object and contains an assay matrix, which contains expression data over a set of
samples for a group of genes, and a string representing the name of a chosen solver.
}
\section{Slots}{

\describe{
\item{\code{mtx.assay}}{An assay matrix of gene expression data}

\item{\code{solver}}{A string matching the designated solver for relating a target gene to transcription factors. 
TReNA currently supports 9 solver strings (default = "lasso"):
\itemize{
\item{\link[=solve.Lasso]{"lasso"}}
\item{\link[=solve.Ridge]{"ridge"}}
\item{\link[=solve.RandomForest]{"randomForest"}}
\item{\link[=solve.BayesSpike]{"bayesSpike"}}
\item{\link[=solve.SqrtLasso]{"sqrtlasso"}}
\item{\link[=solve.LassoPV]{"lassopv"}}
\item{\link[=solve.Pearson]{"pearson"}}
\item{\link[=solve.Spearman]{"spearman"}}
\item{\link[=solve.Ensemble]{"ensemble"}}
}}

\item{\code{quiet}}{A logical denoting whether or not the TReNA object should print output}
}}

\seealso{
\code{\link{solve}}
}
