% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TReNA.R
\docType{class}
\name{TReNA-class}
\alias{TReNA-class}
\alias{.TReNA}
\title{An S4 class to represent a TReNA object}
\arguments{
\item{mtx.assay}{An assay matrix of gene expression data}

\item{solver}{A string matching the designated solver for relating a target gene to transcription factors. (default = "lasso")}
}
\value{
An object of the TReNA class
}
\description{
An S4 class to represent a TReNA object
}
