% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help.R
\docType{package}
\name{TReNA-package}
\alias{TReNA-package}
\title{Inferring Transcriptional Regulation with TReNA}
\description{
TReNA provides a framework for using gene expression data to infer relationships
between a target gene and a set of transcription factors.
}
\seealso{
\code{\link{TReNA}}
}
